Rework mapped read merging to occur before all other bam processing.
This also refactors the pileup and hafpipe steps for bam merging,
as we now have a unified merging step right after mapping.
All optional steps on bam files now operate on the merged files.
This streamlines downstream rules, and ensures that all tools
understand that the merged files are in fact the same sample,
with different units, instead of accidentally treating them as
inpdendent samples (we don't trust all tools to do that correctly).