Update and refine conda envs, to improve stability and solvability.
In particular, the picard and the qualimap envs were long standing
issues, as they did not get solved properly by conda - mamba was needed.
This hopefully helps with #11 and #25 as well.
We updated envs according to the suggestions of @akriese, with the
exception of the vep env, for which we could not easily get the bcftools
version updated without breaking VEP completely.